new_app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.4.2
email-validator==2.1.0.post1
jinja2==3.1.2
uuid==1.30
python-jose==3.3.0
bcrypt==4.0.1 